added test code for client
